requirements/base.txt
Django==4.1.3
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0

celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.4.0

whitenoise==6.2.0

django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0
django-storages==1.13.1

drf-jwt==1.19.2

boto3==1.26.0
attrs==22.1.0

gunicorn==20.1.0
sentry-sdk==1.10.1

names
